+++ adev/shared-docs/components/table-of-contents/table-of-contents.component.ts
@@ -17,7 +17,7 @@ import {
 } from '@angular/core';
 import {Location, ViewportScroller} from '@angular/common';
 import {TableOfContentsLevel} from '../../interfaces/index';
-import {TableOfContentsLoader} from '../../services/table-of-contents-loader.service';
+import {TableOfContentsLoader} from '../../services';
 import {IconComponent} from '../icon/icon.component';
 
 @Component({
