+++ adev/shared-docs/components/viewers/BUILD.bazel
@@ -26,9 +26,12 @@ ng_project(
         "//adev:node_modules/rxjs",
         "//adev/shared-docs/components/breadcrumb",
         "//adev/shared-docs/components/copy-source-code-button",
+        "//adev/shared-docs/components/icon",
         "//adev/shared-docs/components/table-of-contents",
         "//adev/shared-docs/interfaces",
         "//adev/shared-docs/providers",
+        "//adev/shared-docs/services",
+        "//adev/shared-docs/utils",
     ],
 )
