+++ devtools/projects/ng-devtools-backend/src/lib/state-serializer/BUILD.bazel
@@ -9,6 +9,7 @@ ng_project(
         exclude = ["*.spec.ts"],
     ),
     deps = [
+        "//devtools/projects/ng-devtools-backend/src/lib:utils",
         "//devtools/projects/ng-devtools-backend/src/lib/directive-forest",
         "//devtools/projects/protocol",
     ],
