+++ devtools/projects/ng-devtools/src/lib/application-services/BUILD.bazel
@@ -74,6 +74,7 @@ ts_test_library(
         "//:node_modules/@angular/common",
         "//:node_modules/@angular/core",
         "//devtools/projects/ng-devtools/src/lib/application-environment",
+        "//devtools/projects/ng-devtools/src/lib/application-operations",
         "//devtools/projects/ng-devtools/src/lib/application-providers:window",
         "//devtools/projects/ng-devtools/src/lib/application-services/test-utils:app_operations_mock",
         "//devtools/projects/ng-devtools/src/lib/application-services/test-utils:settings_mock",
