+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/directive-explorer/directive-forest/index-forest/index-forest.spec.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {indexForest} from './';
+import {indexForest} from './index';
 import {DevToolsNode} from '../../../../../../../protocol';
 
 describe('indexForest', () => {
