+++ packages/common/upgrade/BUILD.bazel
@@ -13,6 +13,7 @@ ng_project(
         ],
     ),
     deps = [
+        "//:node_modules/rxjs",
         "//packages/common",
         "//packages/core",
         "//packages/upgrade",
