+++ packages/core/test/acceptance/selectorless/BUILD
@@ -1,13 +1,20 @@
-load("//tools:defaults2.bzl", "angular_jasmine_test", "ng_project", "ng_web_test_suite")
+load("//tools:defaults2.bzl", "angular_jasmine_test", "ng_project", "ng_web_test_suite", "ts_config")
 
 package(default_visibility = ["//visibility:private"])
+
+ts_config(
+    name = "tsconfig",
+    src = "tsconfig.json",
+    deps = [
+        "//packages/core:tsconfig_test",
+    ],
+)
 
 ng_project(
     name = "selectorless_test_lib",
     testonly = True,
     srcs = glob(["**/*.ts"]),
-    extends = "//packages/core:tsconfig_test",
-    tsconfig = ":tsconfig.json",
+    tsconfig = ":tsconfig",
     visibility = ["//:__pkg__"],
     deps = [
         "//packages/core",
