+++ packages/upgrade/static/BUILD.bazel
@@ -16,6 +16,7 @@ ng_project(
         "//packages/core",
         "//packages/platform-browser",
         "//packages/upgrade",
+        "//packages/upgrade/src/common",
     ],
 )
