+++ packages/upgrade/static/testing/BUILD.bazel
@@ -14,6 +14,7 @@ ng_project(
     ),
     deps = [
         "//packages/core/testing",
+        "//packages/upgrade/src/common",
         "//packages/upgrade/static",
     ],
 )
