+++ packages/zone.js/test/closure/zone.closure.ts
@@ -5,7 +5,6 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.dev/license
  */
-import '../zone.umd.js';
 const testClosureFunction = () => {
   const logs: string[] = [];
   // call all Zone exposed functions
